pyapp/pyproject.toml
[project]
name = "pyapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "parsa", email = "parsa.miraghaie@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
  "click>=8.1.8",
  "go-task-bin>=3.43.3",
  "litellm>=1.68.0",
  "loguru>=0.7.3",
  "pydantic>=2.11.4",
  "python-dotenv>=1.1.0",
  "serve[mlflow]",
  "taskpy",
  "toml>=0.10.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
pyapp-cmd = "pyapp.cli.main:main"

[dependency-groups]
dev = ["pytest>=8.3.5", "ipykernel>=6.29.5"]

local = ["arize-phoenix-otel>=0.9.2"]

local-langchain-base = [
  { include-group = "local" },
  "openinference-instrumentation-langchain>=0.1.42",
  "langchain-core>=0.3.58",
  "langchain-community>=0.2.1",
]

local-langchain = [{ include-group = "local-langchain-base" }]

local-langchain-embeddings = [
  { include-group = "local-langchain-base" },
  "fastembed>=0.6.1",
]

devserve = [
  { include-group = "local-langchain-embeddings" },
  { include-group = "dev" },
]

[project.optional-dependencies]
simple = ["langchain-core>=0.3.58"]
local-langchain = [
  "arize-phoenix-otel>=0.9.2",
  "openinference-instrumentation-langchain>=0.1.42",
  "langchain-core>=0.3.58",
  "langchain-community>=0.2.1",
]
local-langchain-embeddings = [
  "arize-phoenix-otel>=0.9.2",
  "openinference-instrumentation-langchain>=0.1.42",
  "langchain-core>=0.3.58",
  "langchain-community>=0.2.1",
  "fastembed>=0.6.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.uv.sources]
taskpy = { git = "https://github.com/Parsa-Mir/taskpy.git" }
pyapp = { path = "src/pyapp" }
serve = { path = "../serve", editable = true }
